Added the statement
nfacc=0

to the Xvert.R program in the R directory to initialize the variable nfacc.
 
Also added CONTINUE to end each of the DO loop in the program cnvrt.f which 
is in the src, src-i386, and src-x64 directories. This is due to the fact 
that Rtools42 that uses gcc 10.3 as the compiler where DO loops that don't 
end in END DO or CONTINUE generate a warning that this is a deleted feature
in Fortran when it is compiling cnvrt.f.  

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/daewr@216 edb9625f-4e0d-4859-8d74-9fd3b1da38cb